requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
Flask==2.3.2
heyoo==0.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
protobuf==4.21.12
requests==2.31.0
requests-toolbelt==1.0.0
rivescript==1.15.0
six==1.16.0
typing==3.7.4.3
urllib3==2.0.4
Werkzeug==2.3.7
